trying to push from 'Jing' to 'Amelea'

+++ src/main/java/com/example/thymeleafDemo/Resposity/DeviceRepo.java
@@ -3,5 +3,5 @@ package com.example.thymeleafDemo.Resposity;
 import com.example.thymeleafDemo.Model.Device;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeviceRepo extends JpaRepository<Device,Integer> {
+public interface DeviceRepo extends JpaRepository<Device,Long> {
 }
